🧹 deduplicate replaceImagePlaceholder logic into MarkdownUtils
Extracted the identical `replaceImagePlaceholder` and `applyAltTextToMarkdown` functions from `CreateVoiceActivity.kt` and `DashboardPostDetailActivity.kt` into the shared `MarkdownUtils.kt` object. Updated the call sites in both activities to use the shared implementation.
