+++ test_script.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+echo "Testing replaceImagePlaceholder and applyAltTextToMarkdown behavior"
